% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_2025_excel}
\alias{parse_2025_excel}
\title{Parse 2025 Excel file with multiple sheets}
\usage{
parse_2025_excel(excel_path, sheets)
}
\arguments{
\item{excel_path}{Path to downloaded Excel file}

\item{sheets}{Vector of sheet names}
}
\value{
Combined data frame
}
\description{
The 2025 file has three sheets that need to be combined:
\itemize{
\item "SY25 by School": School-level data
\item "SY25 by District": District-level aggregates
\item "SY25 State": State-level aggregate
}
}
\keyword{internal}
