% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_era2_excel}
\alias{parse_era2_excel}
\title{Parse Era 2 Excel file (2019-present)}
\usage{
parse_era2_excel(excel_path, end_year)
}
\arguments{
\item{excel_path}{Path to downloaded Excel file}

\item{end_year}{School year end}
}
\value{
Data frame with enrollment data
}
\description{
Newer NM PED enrollment files include demographic subgroup breakdowns
by race/ethnicity and special populations.
}
\keyword{internal}
