% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_era1_enr}
\alias{process_era1_enr}
\title{Process Era 1 enrollment data (2016-2018)}
\usage{
process_era1_enr(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Era 1 data only has enrollment by grade, without demographic subgroups.
}
\keyword{internal}
