% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_era1_rows}
\alias{process_era1_rows}
\title{Process Era 1 rows into standard format}
\usage{
process_era1_rows(df, end_year, row_type, find_col)
}
\arguments{
\item{df}{Data frame of rows to process}

\item{end_year}{School year end}

\item{row_type}{"School" or "District"}

\item{find_col}{Function to find columns by pattern}
}
\value{
Processed data frame
}
\description{
Process Era 1 rows into standard format
}
\keyword{internal}
