% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_year}
\alias{validate_year}
\title{Validate year parameter}
\usage{
validate_year(end_year, allow_future = FALSE)
}
\arguments{
\item{end_year}{School year end to validate}

\item{allow_future}{If TRUE, allows years up to current + 1}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate year parameter
}
\keyword{internal}
