% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_era2_enr}
\alias{process_era2_enr}
\title{Process Era 2 enrollment data (2019-present)}
\usage{
process_era2_enr(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Era 2 data includes demographic subgroups by race/ethnicity and
special populations.
}
\keyword{internal}
